fix(major refactor): the app is now majorly refactored, where the layer and the tracked stuff (vehicle and geom) are pure view components, and just one level below the Vehicles component, which has all the fetch logic
