+++ packages/vehicles/src/VehicleLayer/index.js
@@ -4,7 +4,6 @@ import { FeatureGroup } from "react-leaflet";
 
 import { vehicleType } from "../types";
 import VehicleMarker from "../VehicleMarker";
-import VehicleGeometry from "../VehicleGeometry";
 
 function VehicleLayer(props) {
   const { vehicles } = props;
@@ -34,7 +33,6 @@ function VehicleLayer(props) {
             setTracked={props.setTracked}
           />
         ))}
-      <VehicleGeometry trackedVehicle={trackedVehicle} />
     </FeatureGroup>
   );
 }
